requirements.txt
cryptography==38.0.1
Flask==2.2.2
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
mysql-connector==2.2.9
mysql-connector-python==8.0.32
pathlib==1.0.1
pycryptodome==3.17
PyJWT==2.6.0
python-dotenv==0.21.1
gunicorn==20.1.0